ui: remove rolled out flags   * ref https://github.com/cern-sis/issues-inspire/issues/734
