+++ ui/public/config.js
@@ -6,11 +6,6 @@
     REACT_APP_SENTRY_ENVIRONMENT: null,
     BANNERS: null,
     DISPLAY_GUIDE_ON_START: false,
-    ASSIGN_OWN_PROFILE_UI_FEATURE_FLAG: true,
-    ASSIGN_DIFFERENT_PROFILE_UI_FEATURE_FLAG: true,
-    ASSIGN_NO_PROFILE_UI_FEATURE_FLAG: true,
-    ASSIGN_NOT_LOGGED_IN_FEATURE_FLAG: true,
-    SELF_CURATION_BUTTON: true,
     INSPIRE_WORKFLOWS_DAGS_URL: null,
     BACKOFFICE_URL: 'http://localhost:8001',
     EDITOR_SCHEMA_URL: 'https://inspirebeta.net/schemas/records/authors.json',
